pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
